Refactor : 가시성이 좋게 Request와 Response로 구분하여 패키지 변경
