+++ src/main/java/model/http/response/ResponseHeaders.java
@@ -1,4 +1,6 @@
-package model.http;
+package model.http.response;
+
+import model.http.ContentType;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
